+++ mayan/apps/common/views.py
@@ -32,6 +32,7 @@ from .generics import (  # NOQA
 from .icons import icon_setup
 from .menus import menu_tools, menu_setup
 from .permissions_runtime import permission_error_log_view
+from .settings import setting_home_view
 from .utils import check_version
 
 
@@ -243,7 +244,7 @@ class PackagesLicensesView(SimpleView):
 
 
 class RootView(SimpleView):
-    extra_context = {'home_view': settings.HOME_VIEW}
+    extra_context = {'home_view': setting_home_view.value}
     template_name = 'appearance/root.html'